plugins are now instantiated in the initialization function
