+++ src/plugins/candlestick_color/__init__.py
@@ -1,9 +1,7 @@
 """Plugin for candlestick colors."""
-from StockBench.plugin.plugin import PluginManager
 from .candlestick_color import CandlestickColorPlugin
 
 
 def initialize():
-    PluginManager.register('candlestick_color_plugin', CandlestickColorPlugin)
-    return {'candlestick_color_plugin': CandlestickColorPlugin}
+    return {'candlestick_color_plugin': CandlestickColorPlugin()}
