+++ src/plugins/price/__init__.py
@@ -1,8 +1,6 @@
 """Plugin for price."""
-from StockBench.plugin.plugin import PluginManager
 from .price import PricePlugin
 
 
 def initialize():
-    PluginManager.register('price_plugin', PricePlugin)
-    return {'price_plugin': PricePlugin}
+    return {'price_plugin': PricePlugin()}
