+++ src/plugins/stop_profit/__init__.py
@@ -1,8 +1,6 @@
 """Plugin for the stop profit"""
-from StockBench.plugin.plugin import PluginManager
 from .stop_profit import StopProfitPlugin
 
 
 def initialize():
-    PluginManager.register('stop_profit_plugin', StopProfitPlugin)
-    return {'stop_profit_plugin': StopProfitPlugin}
+    return {'stop_profit_plugin': StopProfitPlugin()}
