+++ src/plugins/volume/__init__.py
@@ -1,8 +1,6 @@
 """Plugin for volume."""
-from StockBench.plugin.plugin import PluginManager
 from .volume import VolumePlugin
 
 
 def initialize():
-    PluginManager.register('volume_plugin', VolumePlugin)
-    return {'volume_plugin': VolumePlugin}
+    return {'volume_plugin': VolumePlugin()}
